- Fix logout problem - Fix local connection not established - Fix feed tabs not showing up torrent client's items - Remove unnecessary mehtods - (probably) fix fonts not appear properly in the Store version

+++ Cafeine/Cafeine/Design/RemoteTorrent/TorrentModel.cs
@@ -8,9 +8,9 @@ namespace Cafeine.Design.RemoteTorrent {
     public class TorrentModel {
         public string hash;
         public string name;
-        public int size;
+        public long size;
         public float progress;
-        public int eta;
+        public long eta;
         public string state;
     }
 }
